Aplicação de DTO

+++ src/main/java/com/programacao/web/fatec/api_fatec/domain/ContasPagarRepository.java
@@ -2,6 +2,6 @@ package com.programacao.web.fatec.api_fatec.domain;
 import com.programacao.web.fatec.api_fatec.entities.ContasPagar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContasPagarRepository extends JpaRepository<ContasPagar, Integer> {
+public interface ContasPagarRepository extends JpaRepository<ContasPagar, Long> {
     
 }
