+++ src/main/java/com/programacao/web/fatec/api_fatec/domain/ContasReceberRepository.java
@@ -2,6 +2,6 @@ package com.programacao.web.fatec.api_fatec.domain;
 import com.programacao.web.fatec.api_fatec.entities.ContasReceber;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContasReceberRepository extends JpaRepository<ContasReceber, Integer> {
+public interface ContasReceberRepository extends JpaRepository<ContasReceber, Long> {
     
 }
